08_AI_agent/main.py
import os
import argparse
from dotenv import load_dotenv
from google import genai
from google.genai import types
from prompts import system_prompt
from call_function import available_functions, call_function


def main():
    # Setup
    parser = argparse.ArgumentParser(description="chatbot")
    parser.add_argument("user_prompt", type=str, help="User prompt")
    parser.add_argument("--verbose", action="store_true", help="Enable verbose output")
    args = parser.parse_args()

    load_dotenv()
    api_key = os.environ.get("GEMINI_API_KEY")
    if not api_key:
        raise RuntimeError("API key not found check .env file.")
 
    client = genai.Client(api_key=api_key)
    messages = [types.Content(role="user", parts=[types.Part(text=args.user_prompt)])]
    for i in range(20):
        final_result = generate_content(client, messages, args)
        if final_result:
            print(final_result)
            return
    print("Maximum iterations reached")
    sys.exit(1)



def generate_content(client, messages, args):
    response = client.models.generate_content(
        model="gemini-2.5-flash",
        contents=messages,
        config=types.GenerateContentConfig(
            tools=[available_functions],
            system_instruction=system_prompt
        ),
    )

    if not response.usage_metadata:
        raise RuntimeError("No usage metadata found, verify request.")

    if response.candidates:
        for candidate in response.candidates:
            if candidate.content:
                messages.append(candidate.content)

    if response.function_calls == None:
        if args.verbose:
            print(f"User prompt: {args.user_prompt}")
            print(f"Prompt tokens: {response.usage_metadata.prompt_token_count}")
            print(f"Response tokens: {response.usage_metadata.candidates_token_count}")
            print(response.text)
        else:
            print(response.text)
        return response.text

    if response.function_calls:
        function_calls = response.function_calls
        function_results = []
        for function_call in function_calls:
            function_call_result = call_function(function_call)
            if not function_call_result.parts:
                raise Exception("Function call .parts is empty!!")
            if function_call_result.parts[0].function_response == None:
                raise Exception("Function call .parts.function_response == None")
            if function_call_result.parts[0].function_response.response == None:
                raise Exception("Function call .parts.function_response.response == None")
            function_results.append(function_call_result.parts[0])
        if args.verbose:
            print(f"-> {function_call_result.parts[0].function_response.response}")
        messages.append(types.Content(role="user", parts=function_results))

if __name__ == "__main__":
    main()
